Refatoração e melhorias no código: - Adicionada configuração do arquivo .env para armazenar variáveis de ambiente. - Configuração do Flask para carregar automaticamente variáveis do .env utilizando python-dotenv. - Refatoração das funções principais para maior modularidade e clareza:   - Funções de processamento de texto separadas para maior legibilidade e reutilização.   - Particionamento de funções de conexão com a API e manipulação de respostas. - Instalação das dependências necessárias no requirements.txt, incluindo Flask, spaCy, python-dotenv, requests e outras bibliotecas. - Adicionado aviso no requirements.txt sobre a compatibilidade do spaCy com Python até a versão 3.12. - Garantido que o código esteja compatível com as versões recentes de bibliotecas, como o spaCy, e que o ambiente seja configurado corretamente.
