+++ .env
@@ -1,0 +1,2 @@
+IMAGES_PATH=data/images
+INTENTS_PATH=data/intents.json
