+++ README.md
@@ -1,1 +1,21 @@
-# Pract.Chatbot
+# Pract Chatbot
+
+
+```
+chatbot-erp/
+│
+├── api/
+│   ├── __init__.py
+│   ├── main.py          # Ponto de entrada da API
+│   ├── models.py        # Definições de modelos Pydantic
+│   ├── utils.py         # Funções utilitárias (processamento de texto, imagens, etc.)
+│   ├── intents.py       # Lógica para carregar e gerenciar intenções
+│   └── responses.py     # Lógica para gerar respostas
+│
+├── data/
+│   ├── intents.json     # Arquivo JSON com as intenções
+│   └── images/          # Pasta com as imagens dos produtos
+│
+├── requirements.txt     # Dependências do projeto
+└── README.md            # Documentação do projeto
+```
