+++ api/models.py
@@ -1,0 +1,4 @@
+from pydantic import BaseModel
+
+class Message(BaseModel):
+    question: str
